♻️ refactor: incapsulate loading state into PostList
