+++ src/shared/lib/hoc/withLoading.module.css
@@ -1,0 +1,3 @@
+.loadingContainer {
+  text-align: center;
+}
