+++ src/widgets/CommentList/ui/CommentList.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/shared/ui/Button/Button";
 import styles from "./CommentList.module.css";
 
 type CommentListProps = {
-  postId: number,
+  postId: number;
 };
 
 export function CommentList({ postId }: CommentListProps) {
@@ -14,29 +14,29 @@ export function CommentList({ postId }: CommentListProps) {
 
   const comments = mockComments.filter((comment) => comment.postId === postId);
 
-  const commentList = useMemo(() => (
-    comments.map((comment) => (
-      <li className={styles.commentItem} key={comment.id}>
-        <p className={styles.commentItemName}>{comment.name}</p>
-        <p className={styles.commentItemBody}>{comment.body}</p>
-      </li>
-    ))
-  ), [mockComments]);
+  const commentList = useMemo(
+    () =>
+      comments.map((comment) => (
+        <li className={styles.commentItem} key={comment.id}>
+          <p className={styles.commentItemName}>{comment.name}</p>
+          <p className={styles.commentItemBody}>{comment.body}</p>
+        </li>
+      )),
+    [mockComments]
+  );
 
   return (
     <div className={styles.commentSection}>
-      <Button 
-        variant="text" 
-        onClick={toggleComments} 
+      <Button
+        variant="text"
+        onClick={toggleComments}
         className={`${styles.btnTrigger} ${isExpanded ? styles.active : ""}`}
       >
-        {isExpanded ? "Свернуть комментарии" : `Показать комментарии (${comments.length})`}
+        {isExpanded
+          ? "Свернуть комментарии"
+          : `Показать комментарии (${comments.length})`}
       </Button>
-      {isExpanded && (
-        <ul className={styles.commentList}>
-          {commentList}
-        </ul>
-      )}
+      {isExpanded && <ul className={styles.commentList}>{commentList}</ul>}
     </div>
-  )
+  );
 }
